Refactor: 코드 재배치
- dataSouce 에서의 데이터 가공 로직 repository 로 이동
- map 함수 적용
