+++ android/app/src/main/java/com/example/todo/respository/ActionLogDataSource.kt
@@ -1,7 +1,8 @@
 package com.example.todo.respository
 
 import com.example.todo.model.ActionLog
+import com.example.todo.network.JsonActionLog
 
 interface ActionLogDataSource {
-    suspend fun getActionLogs(): List<ActionLog>
+    suspend fun getActionLogs(): List<JsonActionLog>
 }
