+++ android/app/src/main/java/com/example/todo/respository/ActionLogRemoteDataSource.kt
@@ -11,46 +11,8 @@ import com.example.todo.network.RetrofitClient
 
 class ActionLogRemoteDataSource(
     private val retrofitClient: RetrofitClient,
-    private val context: Context
 ) : ActionLogDataSource {
-    override suspend fun getActionLogs(): List<ActionLog> {
-        val response = retrofitClient.getActionLog()
-        val resultList = mutableListOf<ActionLog>()
-        response.histories.forEach {
-            val actionLog = stringActionLogToActionLog(it)
-            resultList.add(actionLog)
-        }
-        return resultList
-    }
-
-    private fun stringActionLogToActionLog(jsonActionLog: JsonActionLog): ActionLog {
-        val actionType = stringActionTypeToEnum(jsonActionLog.actionType)
-        val pastProgressType = stringProgressTypeToEnum(jsonActionLog.pastLocation)
-        val nowProgressType = stringProgressTypeToEnum(jsonActionLog.nowLocation)
-
-        return ActionLog(
-            jsonActionLog.cardTitle,
-            actionType,
-            jsonActionLog.historyDate,
-            pastProgressType,
-            nowProgressType
-        )
-    }
-
-    private fun stringProgressTypeToEnum(stringProgressType: String): ProgressType {
-        return when (stringProgressType) {
-            context.getString(R.string.progress_todo) -> ProgressType.TO_DO
-            context.getString(R.string.progress_ing) -> ProgressType.IN_PROGRESS
-            else -> ProgressType.DONE
-        }
-    }
-
-    private fun stringActionTypeToEnum(stringActionType: String): ActionType {
-        return when (stringActionType) {
-            context.getString(R.string.action_add) -> ActionType.ADD
-            context.getString(R.string.action_move) -> ActionType.MOVE
-            context.getString(R.string.action_remove) -> ActionType.REMOVE
-            else -> ActionType.UPDATE
-        }
+    override suspend fun getActionLogs(): List<JsonActionLog> {
+       return retrofitClient.getActionLog().histories
     }
 }
